+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortToolMaker.java
@@ -17,5 +17,9 @@ public class AIPortToolMaker {
     public String name;
     public String tags;
     public long agentId;
+    public int agentStatus;
+    public String agentName;
     public long userId;
+    public String userName;
+    public String userAccount;
 }